plant_disease/plant_project/requirements.txt
absl-py==0.9.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
bleach==1.5.0
bz2file==0.98
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
dill==0.3.1.1
Django==2.0.10
django-cors-headers==2.4.0
django-hosts==3.0
djangorestframework==3.8.2
dopamine-rl==3.0.1
entrypoints==0.3
enum34==1.1.6
Flask==1.1.1
future==0.18.2
gast==0.2.2
gevent==1.4.0
gin-config==0.3.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
googleapis-common-protos==1.51.0
greenlet==0.4.15
grpcio==1.27.2
gunicorn==19.8.1
gym==0.16.0
h5py==2.10.0
html5lib==0.9999999
httplib2==0.17.0
idna==2.8
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.2
jupyterlab==1.2.6
jupyterlab-server==1.0.6
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kfac==0.2.0
kiwisolver==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mesh-tensorflow==0.1.9
mistune==0.8.4
mpmath==1.1.0
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.2.0.32
opt-einsum==3.1.0
pandocfilters==1.4.2
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==5.1.0
pipenv==2018.7.1
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==3.0.3
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.4.10
Pygments==2.5.2
pyparsing==2.4.6
pypng==0.0.20
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
pyzmq==18.1.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
setuptools==45.2.0
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
sympy==1.5.1
tb-nightly==1.15.0a20190821
tensor2tensor==1.14.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-cpu==1.15.0rc0
tensorflow-datasets==2.0.0
tensorflow-estimator==1.14.0
tensorflow-gan==2.0.0
tensorflow-gpu==1.15.0
tensorflow-hub==0.7.0
tensorflow-metadata==0.21.1
tensorflow-probability==0.7.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tf-estimator-nightly==1.14.0.dev2019090401
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==16.0.0
virtualenv-clone==0.3.0
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
wrapt==1.12.0
zipp==2.2.0
